requirements.txt
# Core AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# AgentCore and Strands
bedrock-agentcore-starter-toolkit>=0.1.0
strands-agents>=0.1.0

# FastMCP for agent communication
fastmcp>=0.1.0

# Data validation and models
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database and API
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# Web framework for Lambda
fastapi>=0.108.0
mangum>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0
moto>=4.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
email-validator>=2.1.0
python-multipart>=0.0.6

# Development tools
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0